+++ .checkmarx/checkmarxIgnoredTempList.json
@@ -1,0 +1,7 @@
+[
+  {
+    "FileName": "result.go",
+    "Line": 2630,
+    "RuleID": 3007
+  }
+]
